# api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: CleanAUS Enterprise API
  description: |
    Enterprise-grade cleaning services platform for Australia.
    Supports residential, commercial, and strata cleaning services across all Australian regions.
  version: 1.0.0
  contact:
    name: CleanAUS API Support
    email: api@cleanaus.com.au
    url: https://api.cleanaus.com.au
  license:
    name: Proprietary
    url: https://cleanaus.com.au/license
servers:
  - url: https://api.cleanaus.com.au/v1
    description: Production
  - url: https://staging-api.cleanaus.com.au/v1
    description: Staging
  - url: http://localhost:3001/v1
    description: Local Development
...
paths:
...
  /payments/create-intent:
    post:
      summary: Create Stripe payment intent
      tags: [Payments]
      operationId: createPaymentIntent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentIntentRequest'
      responses:
        '200':
          description: Payment intent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntentResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentIntentRequest:
      type: object
      required: [bookingId, amount]
      properties:
        bookingId:
          type: string
          format: uuid
        amount:
          type: number
          format: float
          description: Amount in cents AUD
        paymentMethod:
          type: string
          enum: [card, bank_transfer, direct_debit]
        saveCard:
          type: boolean
          default: false
    PaymentIntentResponse:
      type: object
      required: [clientSecret, paymentIntentId, amount, currency]
      properties:
        clientSecret:
          type: string
          description: Stripe client secret for payment confirmation
        paymentIntentId:
          type: string
          description: Stripe payment intent ID
        amount:
          type: integer
          description: Amount in cents AUD
        currency:
          type: string
          default: aud
          description: Currency code (AUD)
...